add layers.py
